Revised Makefile and runmpich to work on PC with new directory structure.
Made a couple of casts to eliminate compiler warnings.

project file changes because of compilations.

git-svn-id: svn://136.177.114.72/svn_GW/phastpp/trunk@3381 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/phastinput/KDtree/KDtree.cpp
@@ -20,7 +20,7 @@ v(t.v)
 KDtree::KDtree(std::vector < Point > &pts):realdata(extents[pts.size()][3])
 {
 	int
-		n = pts.size();
+		n = (int) pts.size();
 	int
 		i,
 		j;
